fix(embed): reject a non-finite camera segment, expose select/isolate on the SDK
`Number('Infinity')` is not `NaN`, so `?camera=Infinity,0` cleared the
`!isNaN` filter and reached `setCameraRotation` with a non-finite azimuth
instead of falling back to home framing. The blank-segment guard does not
catch it either -- the segment is not empty. The acceptance test is now
finiteness, which still admits the all-zero pose the SDK ships.

`EmbedOptions` gains `select` and `isolate`. Both are already in
`EmbedUrlParams` and applied by the viewer once the first model is on
screen, so an initial selection was reachable only by hand-writing the
iframe URL.

Adding two options put `embed-sdk/src/index.ts` at 408 lines, over the
400-line house limit, so the options type and its query-string
serialisation move to `src/options.ts`; `EmbedOptions` is re-exported from
the entry point, so the public surface is unchanged.
