+++ .changeset/embed-url-params-2934.md
@@ -16,3 +16,12 @@ PascalCase.
 `select` and `isolate` now ignore empty or non-positive id segments (e.g. a
 bare `?isolate=,`) instead of treating them as express id `0`, which used to
 isolate nothing and blank the whole model with no error.
+
+`?camera=` now requires every segment to be finite: `Number('Infinity')` is not
+`NaN`, so `?camera=Infinity,0` used to clear the numeric filter and steer the
+view to a non-finite azimuth instead of falling back to home framing.
+
+`EmbedOptions` now carries `select` and `isolate`, serialised as comma-separated
+query parameters. Both were already in `EmbedUrlParams` and applied by the
+viewer, so an initial selection was previously reachable only by hand-writing
+the iframe URL.
